+++ lib/propcheck/utils.ex
@@ -1,0 +1,50 @@
+defmodule PropCheck.Utils do
+  @moduledoc false
+
+  # Merge local options for `PropCheck.quickcheck/2` and
+  # `PropCheck.check/2` with global options. Global options
+  # take precedence.
+  def merge_global_opts(local_opts) do
+    case Application.get_env(:propcheck, :global_verbose) do
+      true ->
+        [:verbose | local_opts]
+      false ->
+        [:quiet | local_opts]
+      nil ->
+        local_opts
+    end
+  end
+
+  # Merge options
+  def merge(opts1, opts2) do
+    opts1
+    |> Enum.concat(opts2)
+    |> Enum.uniq()
+  end
+
+  # Store options in the process dictionary for later retrieval.
+  def put_opts(opts) do
+    Process.put(:property_opts, opts)
+    opts
+  end
+
+  # Retrieve stored options from the process dictionary.
+  def get_opts do
+    Process.get(:property_opts, []) || []
+  end
+
+  # Check if verbose should be enabled
+  def verbose?(opts) do
+    verbose_index = Enum.find_index(opts, & &1 == :verbose)
+    if verbose_index do
+      quiet_index = Enum.find_index(opts, & &1 == :quiet)
+      if quiet_index do
+        verbose_index < quiet_index
+      else
+        true
+      end
+    else
+      false
+    end
+  end
+end