+++ lib/properties.ex
@@ -109,13 +109,14 @@ defmodule PropCheck.Properties do
           tags = [[failing_prop: tag_property({module, name, []})]]
           prop_name = ExUnit.Case.register_test(__ENV__, :property, name, tags)
           def unquote(prop_name)(unquote(var)) do
-            merged_opts = PropCheck.Properties.merge_opts(unquote(opts), unquote(module_default_opts))
+            merged_opts =
+              PropCheck.Properties.merge_opts(unquote(opts), unquote(module_default_opts))
+              |> PropCheck.Utils.merge_global_opts()
+              |> PropCheck.Utils.put_opts()
 
-            # Store opts in process dictionary to make them available within macros such
-            # as forall
-            Process.put(:property_opts, merged_opts)
             p = unquote(block)
             mfa = {unquote(module), unquote(prop_name), []}
+
             execute_property(p, mfa, merged_opts, unquote(store_counter_example))
             :ok
           end